+++ pavement_parking/src/main.rs
@@ -1,17 +1,19 @@
-use std::collections::HashMap;
 use std::io::BufWriter;
 
 use anyhow::{bail, Result};
 use fs_err::File;
 use gdal::vector::LayerAccess;
 use gdal::Dataset;
-use geo::{
-    Contains, Coord, GeodesicLength, Geometry, Intersects, LineString, MapCoordsInPlace,
-    MultiPolygon, Polygon,
-};
-use geojson::{FeatureCollection, FeatureWriter};
+use geojson::FeatureWriter;
 use indicatif::{ProgressBar, ProgressStyle};
-use rstar::{primitives::GeomWithData, RTree, RTreeObject};
+
+use crate::boundaries::Boundaries;
+use crate::census_areas::CensusAreas;
+use crate::roads::Road;
+
+mod boundaries;
+mod census_areas;
+mod roads;
 
 fn main() -> Result<()> {
     let args: Vec<String> = std::env::args().collect();
@@ -19,38 +21,15 @@ fn main() -> Result<()> {
         println!("Call with path to trn_ntwk_roadlink.gpkg");
         std::process::exit(1);
     }
+    let input_path = &args[1];
+    let summaries_output_path = "web/public/summaries.geojson";
+    let pavements_output_path = "web/public/pavements.geojson";
+    let output_areas_output_path = "web/public/output_areas.geojson";
 
-    let boundaries = read_boundaries("inputs/boundaries.geojson")?;
-    let census_areas = read_census_areas("inputs/car_ownership.gpkg")?;
-    gpkg_to_geojson(
-        &args[1],
-        "web/public/summaries.geojson",
-        "web/public/pavements.geojson",
-        "web/public/output_areas.geojson",
-        process_feature,
-        boundaries,
-        census_areas,
-    )
-}
+    let mut boundaries = Boundaries::load("inputs/boundaries.geojson")?;
+    let mut census_areas = CensusAreas::load("inputs/car_ownership.gpkg")?;
 
-fn gpkg_to_geojson<
-    F: Fn(
-        LineString,
-        gdal::vector::Feature,
-        &mut Boundaries,
-        &mut CensusAreas,
-        &mut FeatureWriter<BufWriter<File>>,
-    ) -> Result<()>,
->(
-    input_path: &str,
-    summaries_output_path: &str,
-    pavements_output_path: &str,
-    output_areas_output_path: &str,
-    process: F,
-    mut boundaries: Boundaries,
-    mut census_areas: CensusAreas,
-) -> Result<()> {
-    println!("Processing {input_path}");
+    println!("Reading {input_path}");
     let dataset = Dataset::open(input_path)?;
     // Assume only one layer
     let mut layer = dataset.layer(0)?;
@@ -63,16 +42,7 @@ fn gpkg_to_geojson<
 
     for input_feature in layer.features() {
         progress.inc(1);
-        let mut geo = input_feature.geometry().unwrap().to_geo()?;
-        // Remove unnecessary precision
-        geo.map_coords_in_place(|Coord { x, y }| Coord {
-            x: trim_f64(x),
-            y: trim_f64(y),
-        });
-        let ls: LineString = geo.try_into()?;
-
-        process(
-            ls,
+        handle_road(
             input_feature,
             &mut boundaries,
             &mut census_areas,
@@ -80,90 +50,27 @@ fn gpkg_to_geojson<
         )?;
     }
 
-    // Write all boundaries with non-zero counts
-    let mut summaries_writer =
-        FeatureWriter::from_writer(BufWriter::new(File::create(summaries_output_path)?));
-    for obj in boundaries.rtree.drain() {
-        let mut f = geojson::Feature::from(geojson::Value::from(obj.geom()));
-        let counts = boundaries.counts[&obj.data];
-        if counts[0] + counts[1] + counts[2] > 0 {
-            f.set_property("name", obj.data);
-            f.set_property("red", counts[0]);
-            f.set_property("amber", counts[1]);
-            f.set_property("green", counts[2]);
-            summaries_writer.write_feature(&f)?;
-        }
-    }
+    boundaries.write_output(summaries_output_path)?;
+    census_areas.write_output(output_areas_output_path)?;
 
-    // Write all output_areas with non-zero kerb length
-    let mut oa_writer =
-        FeatureWriter::from_writer(BufWriter::new(File::create(output_areas_output_path)?));
-    for obj in census_areas.rtree.drain() {
-        let info = &census_areas.info[&obj.data];
-        if info.aggregate_kerb_length > 0.0 {
-            let mut f = geojson::Feature::from(geojson::Value::from(obj.geom()));
-            f.set_property("GEO_ID", obj.data);
-            f.set_property("number_of_cars_and_vans", info.number_of_cars_and_vans);
-            f.set_property("aggregate_kerb_length", info.aggregate_kerb_length);
-            f.set_property(
-                "kerb_length_per_car",
-                info.aggregate_kerb_length / info.number_of_cars_and_vans,
-            );
-            oa_writer.write_feature(&f)?;
-        }
-    }
-
-    println!("Wrote '{summaries_output_path}', '{pavements_output_path}' and '{output_areas_output_path}'");
+    println!("\n\nWrote '{summaries_output_path}', '{pavements_output_path}' and '{output_areas_output_path}'");
     Ok(())
 }
 
-fn trim_f64(x: f64) -> f64 {
-    (x * 10e6).round() / 10e6
-}
-
-fn process_feature(
-    geom: LineString,
+fn handle_road(
     input: gdal::vector::Feature,
     boundaries: &mut Boundaries,
     census_areas: &mut CensusAreas,
     writer: &mut FeatureWriter<BufWriter<File>>,
 ) -> Result<()> {
-    let Some(road_average) = input.field_as_double_by_name("roadwidth_average")? else {
-        return Ok(());
-    };
-    let Some(road_minimum) = input.field_as_double_by_name("roadwidth_minimum")? else {
-        return Ok(());
-    };
-    let Some(class) = input.field_as_string_by_name("roadclassification")? else {
+    let Some(road) = Road::new(input)? else {
         return Ok(());
     };
 
-    // Assume that where there are pavements on both sides of the road, then this value is the
-    // sum of both pavements. If there is only one pavement, then this value is the width of that.
-    let Some(pavement_average) =
-        input.field_as_double_by_name("presenceofpavement_averagewidth_m")?
-    else {
-        return Ok(());
-    };
-
-    // Skip roads that shouldn't be analyzed for pavement parking
-    if class == "Motorway" {
-        return Ok(());
-    }
-
-    let direction = match input
-        .field_as_string_by_name("directionality")?
-        .unwrap()
-        .as_str()
-    {
-        "Both Directions" => "both",
-        "In Direction" | "In Opposite Direction" => "one-way",
-        x => bail!("Unknown directionality {x}"),
-    };
-
-    let average_rating_inc_pavements = rating(&class, road_average + pavement_average)?;
-    let average_rating_exc_pavements = rating(&class, road_average)?;
-    let minimum_rating = rating(&class, road_minimum)?;
+    let average_rating_inc_pavements =
+        rating(&road.class, road.road_average + road.pavement_average)?;
+    let average_rating_exc_pavements = rating(&road.class, road.road_average)?;
+    let minimum_rating = rating(&road.class, road.road_minimum)?;
 
     let rating_change = if average_rating_inc_pavements == average_rating_exc_pavements {
         "no_change"
@@ -171,36 +78,20 @@ fn process_feature(
         average_rating_exc_pavements
     };
 
-    let (output_area_geoid, parkable_length) =
-        aggregate_kerb_length_per_oa(census_areas, &geom, &average_rating_exc_pavements, &class)?;
+    let (output_area_geoid, parkable_length) = census_areas.aggregate_kerb_length_per_oa(
+        &road.geom,
+        &average_rating_exc_pavements,
+        &road.class,
+    )?;
 
-    // Find all matching boundaries
-    for obj in boundaries
-        .rtree
-        .locate_in_envelope_intersecting(&geom.envelope())
-    {
-        // TODO Or even just intersects, to handle boundaries?
-        if obj.geom().contains(&geom) {
-            let count = boundaries.counts.get_mut(&obj.data).unwrap();
-            // TODO Use average_rating_exc_pavements for now
-            if average_rating_exc_pavements == "red" {
-                count[0] += 1;
-            } else if average_rating_exc_pavements == "amber" {
-                count[1] += 1;
-            } else if average_rating_exc_pavements == "green" {
-                count[2] += 1;
-            } else {
-                // No change in rating
-                count[3] += 1;
-            }
-        }
-    }
+    // TODO Use average_rating_exc_pavements for now
+    boundaries.handle_road(&road.geom, average_rating_exc_pavements);
 
     // Include the road in the output
-    let mut output_line = geojson::Feature::from(geojson::Value::from(&geom));
-    output_line.set_property("average_width", road_average);
-    output_line.set_property("minimum_width", road_minimum);
-    output_line.set_property("pavement_average_width", pavement_average);
+    let mut output_line = geojson::Feature::from(geojson::Value::from(&road.geom));
+    output_line.set_property("average_width", road.road_average);
+    output_line.set_property("minimum_width", road.road_minimum);
+    output_line.set_property("pavement_average_width", road.pavement_average);
     output_line.set_property("average_rating", average_rating_exc_pavements);
     output_line.set_property("average_rating_inc_pavements", average_rating_inc_pavements);
     output_line.set_property("minimum_rating", minimum_rating);
@@ -210,8 +101,8 @@ fn process_feature(
         output_area_geoid.unwrap_or("NONE".to_string()),
     );
     output_line.set_property("rating_change", rating_change);
-    output_line.set_property("class", class);
-    output_line.set_property("direction", direction);
+    output_line.set_property("class", road.class);
+    output_line.set_property("direction", road.direction);
     writer.write_feature(&output_line)?;
 
     Ok(())
@@ -244,162 +135,4 @@ fn rating(class: &str, width: f64) -> Result<&'static str> {
 
         _ => bail!("Unknown roadclassification {class}"),
     }
-}
-
-fn parkable_kerb_length(geom: &LineString, rating: &str, class: &str) -> f64 {
-    // Returns the length of the kerb where it is possible to park a car
-    // i.e. not on a junction or a pedestrian crossing, etc.
-    // This attempts to implement the table of proposed interventions in
-    // the Pavement Parking Assessment Document
-
-    // TODO Haversine?
-    let raw_length = geom.geodesic_length();
-
-    let kerb_length = match rating {
-        // If the road is wide enough, assume that both sides are parkable
-        "green" => 2.0 * raw_length,
-        "amber" => raw_length,
-        "red" => match class {
-            "A Road" => 0.0,
-            "B Road" | "Classified Unnumbered" | "Unclassified" => raw_length,
-            "Unknown" | "Not Classified" => raw_length,
-            _ => panic!("Unknown class {class}"),
-        },
-        "TODO" => raw_length,
-        _ => panic!("Unknown rating {rating}"),
-    };
-
-    // TODO - additional considerations:
-    // - One-way roads. Is the widths/rating relationship the same as for two-way roads?
-    // - Roads with parking restrictions (including residence parking). How to handle these?
-    // - Maybe subtract a length near each junction, pedestrian crossing, school entrance, etc.
-    // - For short roads, would the same intervention be applied as for long roads?
-
-    kerb_length
-}
-
-// Returns the GEOID of the census area the road is assigned to, and the parkable length. Also
-// updates the census_areas with the kerb length per assigned road segment
-fn aggregate_kerb_length_per_oa(
-    census_area: &mut CensusAreas,
-    geom: &LineString,
-    average_rating: &str,
-    road_class: &str,
-) -> Result<(Option<String>, f64)> {
-    // For each output area, sum the kerb length where it is possible to park a car.
-    // Calculate the parkable kerb length per car in the area.
-
-    // Estimate the length of the kerb where it is possible to park a car
-    let parkable_length = parkable_kerb_length(geom, average_rating, road_class);
-
-    // Assign each road to exactly one output area. If it intersects multiple output areas,
-    // it can be assigned by arbitrary but repeatable method.
-    // For reproducibility, find all of the of the output areas which intersect the road segment
-    // Then select the one with the alphabetically first GEOID.
-    let target_geoid = census_area
-        .rtree
-        .locate_in_envelope_intersecting_mut(&geom.envelope())
-        .filter(|oa| oa.geom().intersects(geom))
-        .map(|oa| oa.data.clone())
-        .min();
-    if let Some(ref geoid) = target_geoid {
-        census_area
-            .info
-            .get_mut(geoid)
-            .unwrap()
-            .aggregate_kerb_length += parkable_length;
-    }
-
-    Ok((target_geoid, parkable_length))
-}
-
-struct Boundaries {
-    rtree: RTree<GeomWithData<Polygon, String>>,
-    // Per boundary name, the count for [red, amber, green]
-    counts: HashMap<String, [usize; 4]>,
-}
-
-struct CensusAreas {
-    rtree: RTree<GeomWithData<Polygon, String>>,
-    info: HashMap<String, CensusArea>,
-}
-
-struct CensusArea {
-    number_of_cars_and_vans: f64,
-    aggregate_kerb_length: f64,
-}
-
-fn read_boundaries(path: &str) -> Result<Boundaries> {
-    println!("Reading boundaries from {path}");
-    let gj: FeatureCollection = fs_err::read_to_string(path)?.parse()?;
-    let mut boundaries = Vec::new();
-    let mut counts = HashMap::new();
-    for f in gj.features {
-        let name = f.property("name").unwrap().as_str().unwrap().to_string();
-        let mp: MultiPolygon = if matches!(
-            f.geometry.as_ref().unwrap().value,
-            geojson::Value::Polygon(_)
-        ) {
-            MultiPolygon(vec![f.try_into()?])
-        } else {
-            f.try_into()?
-        };
-        // MultiPolygon isn't supported, so just insert multiple names
-        for polygon in mp {
-            boundaries.push(GeomWithData::new(polygon, name.clone()));
-        }
-        counts.insert(name, [0, 0, 0, 0]);
-    }
-    println!("Building RTree of {} boundaries", boundaries.len());
-    let rtree = RTree::bulk_load(boundaries);
-
-    Ok(Boundaries { rtree, counts })
-}
-
-fn read_census_areas(path: &str) -> Result<CensusAreas> {
-    println!("Reading census areas from {path}");
-
-    let dataset = Dataset::open(path)?;
-    // Assume only one layer
-    let mut layer = dataset.layer(0)?;
-
-    let progress = ProgressBar::new(layer.feature_count()).with_style(ProgressStyle::with_template(
-        "[{elapsed_precise}] [{wide_bar:.cyan/blue}] {human_pos}/{human_len} ({per_sec}, {eta})").unwrap());
-    let mut output_areas = Vec::new();
-    let mut info = HashMap::new();
-
-    for input_feature in layer.features() {
-        progress.inc(1);
-        let geo = input_feature.geometry().unwrap().to_geo()?;
-        let mp = match geo {
-            Geometry::Polygon(p) => MultiPolygon(vec![p]),
-            Geometry::MultiPolygon(mp) => mp,
-            _ => bail!("Unexpected geometry type"),
-        };
-        let Some(name) = input_feature.field_as_string_by_name("GEO_ID")? else {
-            bail!("Missing GEO_ID");
-        };
-        let Some(number_of_cars_and_vans) = input_feature
-            .field_as_double_by_name("Number of cars or vans: Total: All households")?
-        else {
-            bail!("Missing number_of_cars_and_vans");
-        };
-
-        // MultiPolygon isn't supported, so just insert multiple names
-        for polygon in mp {
-            output_areas.push(GeomWithData::new(polygon, name.clone()));
-        }
-
-        info.insert(
-            name,
-            CensusArea {
-                number_of_cars_and_vans,
-                aggregate_kerb_length: 0.0,
-            },
-        );
-    }
-    println!("Building RTree of {} output areas", output_areas.len());
-    let rtree = RTree::bulk_load(output_areas);
-
-    Ok(CensusAreas { rtree, info })
 }
