+++ pavement_parking/src/roads.rs
@@ -1,0 +1,71 @@
+use anyhow::{bail, Result};
+use geo::{Coord, LineString, MapCoordsInPlace};
+
+pub struct Road {
+    pub geom: LineString,
+    pub class: String,
+    pub road_average: f64,
+    pub road_minimum: f64,
+    /// Assume that where there are pavements on both sides of the road, then this value is the sum
+    /// of both pavements. If there is only one pavement, then this value is the width of that.
+    pub pavement_average: f64,
+    pub direction: String,
+}
+
+impl Road {
+    /// Parse data about one road from the input gpkg. `None` means to skip this road.
+    pub fn new(input: gdal::vector::Feature) -> Result<Option<Self>> {
+        let Some(class) = input.field_as_string_by_name("roadclassification")? else {
+            bail!("Missing roadclassification");
+        };
+        // Skip roads that shouldn't be analyzed for pavement parking
+        if class == "Motorway" {
+            return Ok(None);
+        }
+
+        let mut geom: LineString = input.geometry().unwrap().to_geo()?.try_into()?;
+        // Remove unnecessary precision
+        geom.map_coords_in_place(|Coord { x, y }| Coord {
+            x: trim_f64(x),
+            y: trim_f64(y),
+        });
+
+        let Some(road_average) = input.field_as_double_by_name("roadwidth_average")? else {
+            // Sometimes this really is missing
+            return Ok(None);
+        };
+        let Some(road_minimum) = input.field_as_double_by_name("roadwidth_minimum")? else {
+            // Sometimes this really is missing
+            return Ok(None);
+        };
+
+        let Some(pavement_average) =
+            input.field_as_double_by_name("presenceofpavement_averagewidth_m")?
+        else {
+            bail!("Missing presenceofpavement_averagewidth_m");
+        };
+
+        let direction = match input
+            .field_as_string_by_name("directionality")?
+            .unwrap()
+            .as_str()
+        {
+            "Both Directions" => "both".to_string(),
+            "In Direction" | "In Opposite Direction" => "one-way".to_string(),
+            x => bail!("Unknown directionality {x}"),
+        };
+
+        Ok(Some(Self {
+            geom,
+            class,
+            road_average,
+            road_minimum,
+            pavement_average,
+            direction,
+        }))
+    }
+}
+
+fn trim_f64(x: f64) -> f64 {
+    (x * 10e6).round() / 10e6
+}
